Do not reuse netlink connections
This updates the netlink reader to not reuse netlink connections, which
appears to work around the issue we've observed in cAdvisor where
cAdvisor receives a message from the Kernel it doesn't expect and
proceeds to allocate 4GB of memory for it.

This *also* throws an error (and a log message) in case such a message
shows up, though I haven't seen any so far.